refactor(server): split subscription query and command flows
