+++ apps/server/src/domain/subscriptions/index.ts
@@ -1,8 +1,14 @@
 export * from "./domain-errors";
 export * from "./models";
-export * from "./repository/subscription.repository";
+export * from "./repository/subscription.repository.types";
+export * from "./repository/subscription-command.repository";
+export * from "./repository/subscription-query.repository";
 export * from "./services/activate-subscription.service";
 export * from "./services/create-subscription.service";
 export * from "./services/subscription-flows.shared";
-export * from "./services/subscription.service";
+export * from "./services/subscription.service.types";
+export * from "./services/subscription-command.live";
+export * from "./services/subscription-command.service";
+export * from "./services/subscription-query.live";
+export * from "./services/subscription-query.service";
 export * from "./services/use-subscription-once.service";
