+++ apps/server/src/domain/subscriptions/package-config.ts
@@ -1,5 +1,9 @@
 import type { SubscriptionPackage } from "generated/prisma/client";
 
+/**
+ * Cấu hình runtime cho từng gói subscription.
+ * Giá và số lượt hiện được giữ cứng trong code vì chưa có yêu cầu quản trị động.
+ */
 export type SubscriptionPackageConfig = {
   readonly packageName: SubscriptionPackage;
   readonly price: bigint;
@@ -28,12 +32,18 @@ const PACKAGE_CONFIG: Record<SubscriptionPackage, SubscriptionPackageConfig> = {
   },
 };
 
+/**
+ * Lấy cấu hình của một package cụ thể.
+ */
 export function getSubscriptionPackageConfig(
   packageName: SubscriptionPackage,
 ): SubscriptionPackageConfig {
   return PACKAGE_CONFIG[packageName];
 }
 
+/**
+ * Trả về toàn bộ package đang được hỗ trợ.
+ */
 export function listSubscriptionPackages(): SubscriptionPackageConfig[] {
   return Object.values(PACKAGE_CONFIG);
 }
