+++ apps/server/src/domain/subscriptions/services/activate-subscription.service.ts
@@ -7,22 +7,34 @@ import {
   SubscriptionNotFound as SubscriptionNotFoundError,
   SubscriptionNotPending as SubscriptionNotPendingError,
 } from "../domain-errors";
-import { SubscriptionServiceTag } from "../services/subscription.service";
+import {
+  SubscriptionCommandServiceTag,
+} from "../services/subscription-command.live";
+import {
+  SubscriptionQueryServiceTag,
+} from "../services/subscription-query.live";
 import { computeExpiresAt } from "./subscription-flows.shared";
 
+/**
+ * Kích hoạt ngay một subscription đang ở trạng thái pending.
+ *
+ * Use case này giữ phần check đọc riêng ra trước khi gọi command service,
+ * để lỗi trả về vẫn rõ ràng khi đọc log hoặc debug job auto-activate.
+ */
 export function activateSubscriptionUseCase(args: {
   subscriptionId: string;
   now?: Date;
 }): Effect.Effect<
   SubscriptionRow,
   ActivateSubscriptionFailure,
-  SubscriptionServiceTag
+  SubscriptionQueryServiceTag | SubscriptionCommandServiceTag
 > {
   return Effect.gen(function* () {
-    const service = yield* SubscriptionServiceTag;
+    const queryService = yield* SubscriptionQueryServiceTag;
+    const commandService = yield* SubscriptionCommandServiceTag;
     const now = args.now ?? new Date();
 
-    const subOpt = yield* service.findById(args.subscriptionId);
+    const subOpt = yield* queryService.getById(args.subscriptionId);
     if (Option.isNone(subOpt)) {
       return yield* Effect.fail(
         new SubscriptionNotFoundError({ subscriptionId: args.subscriptionId }),
@@ -37,7 +49,7 @@ export function activateSubscriptionUseCase(args: {
       );
     }
 
-    const activated = yield* service.activate({
+    const activated = yield* commandService.activate({
       subscriptionId: sub.id,
       activatedAt: now,
       expiresAt: computeExpiresAt(now),
