+++ apps/server/src/domain/subscriptions/services/test/subscription-test-kit.ts
@@ -2,43 +2,64 @@ import { Effect, Layer } from "effect";
 
 import type { PrismaClient, Prisma as PrismaNS } from "generated/prisma/client";
 
-import { makeSubscriptionRepository, SubscriptionRepository } from "@/domain/subscriptions";
+import {
+  makeSubscriptionCommandRepository,
+  makeSubscriptionQueryRepository,
+  SubscriptionCommandRepository,
+  SubscriptionCommandServiceLive,
+  SubscriptionCommandServiceTag,
+  SubscriptionQueryRepository,
+  SubscriptionQueryServiceLive,
+  SubscriptionQueryServiceTag,
+} from "@/domain/subscriptions";
 import { Prisma } from "@/infrastructure/prisma";
 import { runEffectWithLayer } from "@/test/effect/run";
 
-import type { SubscriptionServiceTag } from "../subscription.service";
-
-import { SubscriptionServiceLive } from "../subscription.service";
-
-export type SubscriptionDeps = Prisma | SubscriptionRepository | SubscriptionServiceTag;
+export type SubscriptionDeps
+  = | Prisma
+    | SubscriptionQueryRepository
+    | SubscriptionCommandRepository
+    | SubscriptionQueryServiceTag
+    | SubscriptionCommandServiceTag;
 
 export function makeSubscriptionTestLayer(client: PrismaClient) {
-  const subscriptionRepoLayer = Layer.succeed(
-    SubscriptionRepository,
-    makeSubscriptionRepository(client),
+  const subscriptionQueryRepoLayer = Layer.succeed(
+    SubscriptionQueryRepository,
+    SubscriptionQueryRepository.make(makeSubscriptionQueryRepository(client)),
   );
-  const subscriptionServiceLayer = SubscriptionServiceLive.pipe(
-    Layer.provide(subscriptionRepoLayer),
+  const subscriptionCommandRepoLayer = Layer.succeed(
+    SubscriptionCommandRepository,
+    SubscriptionCommandRepository.make(makeSubscriptionCommandRepository(client)),
+  );
+  const subscriptionQueryServiceLayer = SubscriptionQueryServiceLive.pipe(
+    Layer.provide(subscriptionQueryRepoLayer),
+  );
+  const subscriptionCommandServiceLayer = SubscriptionCommandServiceLive.pipe(
+    Layer.provide(Layer.mergeAll(subscriptionQueryRepoLayer, subscriptionCommandRepoLayer)),
   );
 
   return Layer.mergeAll(
     Layer.succeed(Prisma, Prisma.make({ client })),
-    subscriptionRepoLayer,
-    subscriptionServiceLayer,
+    subscriptionQueryRepoLayer,
+    subscriptionCommandRepoLayer,
+    subscriptionQueryServiceLayer,
+    subscriptionCommandServiceLayer,
   );
 }
 
 export function makeSubscriptionRunners(layer: Layer.Layer<SubscriptionDeps>) {
   return {
-    runWithService<A, E>(eff: Effect.Effect<A, E, SubscriptionServiceTag>) {
+    runWithService<A, E, R extends SubscriptionDeps>(eff: Effect.Effect<A, E, R>) {
       return runEffectWithLayer(eff, layer);
     },
-    runInTxWithService<A, E>(
+    runInTxWithService<A, E, R extends SubscriptionDeps>(
       client: PrismaClient,
-      f: (tx: PrismaNS.TransactionClient) => Effect.Effect<A, E, SubscriptionServiceTag>,
+      f: (tx: PrismaNS.TransactionClient) => Effect.Effect<A, E, R>,
     ) {
       return client.$transaction(async tx =>
-        Effect.runPromise(f(tx).pipe(Effect.provide(layer))),
+        Effect.runPromise(
+          f(tx).pipe(Effect.provide(layer)) as Effect.Effect<A, E, never>,
+        ),
       );
     },
   };
