+++ apps/server/src/worker/commands/subscriptions-expire-sweep-once.ts
@@ -2,9 +2,10 @@ import { Effect } from "effect";
 import process from "node:process";
 
 import {
-  SubscriptionRepositoryLive,
-  SubscriptionServiceLive,
-  SubscriptionServiceTag,
+  SubscriptionCommandRepositoryLive,
+  SubscriptionCommandServiceLive,
+  SubscriptionCommandServiceTag,
+  SubscriptionQueryRepositoryLive,
 } from "@/domain/subscriptions";
 import { PrismaLive } from "@/infrastructure/prisma";
 import logger from "@/lib/logger";
@@ -12,12 +13,13 @@ import logger from "@/lib/logger";
 async function main() {
   await Effect.runPromise(
     Effect.gen(function* () {
-      const service = yield* SubscriptionServiceTag;
+      const service = yield* SubscriptionCommandServiceTag;
       const expiredCount = yield* service.markExpiredNow(new Date());
       return expiredCount;
     }).pipe(
-      Effect.provide(SubscriptionServiceLive),
-      Effect.provide(SubscriptionRepositoryLive),
+      Effect.provide(SubscriptionCommandServiceLive),
+      Effect.provide(SubscriptionQueryRepositoryLive),
+      Effect.provide(SubscriptionCommandRepositoryLive),
       Effect.provide(PrismaLive),
       Effect.tap(expiredCount =>
         Effect.sync(() => {
